// src/app/api/swagger/models/api-purchase-category.ts
/* tslint:disable */
/* eslint-disable */
export enum ApiPurchaseCategoryAltEnum {
  DefenceCountry = 1,
  UseOfAtomicEnergy = 2,
  SphereOfNaturalMonopolies = 3,
  OutsideRussianFederation = 4,
  FinancialServices = 5,
  StateSecret = 6,
  GovernmentDecisionPart16Article4 = 7,
  WaterSupplyServices = 8,
  WorkOfExecutivePower = 9,
  SupervisionOfAuthorServices = 10,
  RealEstate = 11,
  Energy = 12,
  EnergyRecycling = 13,
  Railways = 14,
  IntellectualActivityResults = 15,
  AirTransportationServices = 16,
  LargeDiameterPipes = 17,
  RadioactiveAndChemicalHazardsPruducts = 18,
  MaintenanceServices = 19,
  ProjectsOfShelfProducts = 20,
  MobileRadiotelephoneCommunicationServices = 21,
  EducationalOrganizationsServices = 22,
  MainAndSubsidaryEconomicalSocietyServices = 23,
  SecurityProtectionPersonalInformationServicesGreaterThan200 = 24,
  AuditServicesGreater10 = 25,
  RoughDiamonds = 26,
  PurchasesByStateOwnedCompanies = 27,
  ConstructionAndReconstructionServicesForExtremelyDangerousObjects = 28,
  DesigningConstructionReconsctructionServicesForDangerousObjects = 29
}